Make PauseAfterDownload() work for COEP with a new service worker
Currently Cross-Origin-Embedder-Policy header is parsed in
ServiceWorkerNewScriptLoader when launching a new service worker.
However, the COEP in the response header for the main script isn't
applied for the network loader created on launching a service worker
because the network loader is created before starting to load the main
script.
This CL is a preparation to make that work. PauseAfterDownload() is
going to be used to pause the service worker's script evaluation until
the network loader is updated with the correct COEP value. In the next
patch, SubresourceLoaderUpdater will be used to replace the subresource
loader factory with the one for which the new policy is applied.

Bug: 1039613
Change-Id: Ice82eef86beec2f2f5c5e91f3aa136229f14c3ab